refactor with combineReducers
